[Refactor] Modified the ioctl usage to make it more reusable
