+++ interface_linux.go
@@ -1,9 +1,6 @@
 package socketcan
 
 import (
-	"fmt"
-	"unsafe"
-
 	"golang.org/x/sys/unix"
 )
 
@@ -12,24 +9,6 @@ func getIfIndex(itf Interface, ifName string) (int, error) {
 	if err != nil {
 		return 0, err
 	}
-	if len(ifNameRaw) > unix.IFNAMSIZ {
-		return 0, fmt.Errorf("Maximum ifname length is %d characters", unix.IFNAMSIZ)
-	}
-
-	type ifreq struct {
-		Name  [unix.IFNAMSIZ]byte
-		Index int
-	}
-	var ifReq ifreq
 	fd := itf.SocketFD()
-	copy(ifReq.Name[:], ifNameRaw)
-	_, _, errno := unix.Syscall(unix.SYS_IOCTL,
-		uintptr(fd),
-		unix.SIOCGIFINDEX,
-		uintptr(unsafe.Pointer(&ifReq)))
-	if errno != 0 {
-		return 0, fmt.Errorf("ioctl: %v", errno)
-	}
-
-	return ifReq.Index, nil
+	return IoctlGetIfIndex(fd, ifNameRaw)
 }
